Omstrukturerer filstruktur for ny sykepengesoknad arbeidstaker
- Gjor det litt lettere a finne frem om man ikke kjenner koden godt fra for.
- Gjor det lettere om man vil trekke ut soknad i egen app i fremtiden.
