+++ js/sykepengesoknad/arbeidtaker/NySendtSoknadArbeidstaker.js
@@ -1,18 +1,12 @@
 import React from 'react';
-import {
-    getLedetekst,
-    sykmelding as sykmeldingPt,
-    tilLesbarDatoMedArstall,
-    SykmeldingUtdrag,
-    Utvidbar,
-} from '@navikt/digisyfo-npm';
+import { getLedetekst, sykmelding as sykmeldingPt, SykmeldingUtdrag, tilLesbarDatoMedArstall, Utvidbar, } from '@navikt/digisyfo-npm';
 import { connect } from 'react-redux';
-import Oppsummeringsvisning from '../soknad-felles-oppsummering/Oppsummeringsvisning';
+import Oppsummeringsvisning from '../../components/soknad-felles-oppsummering/Oppsummeringsvisning';
 import { soknad as soknadPt } from '../../propTypes';
-import Soknadtopp from '../soknad-felles/Soknadtopp';
+import Soknadtopp from '../../components/soknad-felles/Soknadtopp';
 import { finnSykmelding } from '../../utils/soknad-felles/soknadSetup';
 import { VAER_KLAR_OVER_AT } from '../../enums/tagtyper';
-import Statuspanel, { StatusNokkelopplysning, Statusopplysninger } from '../Statuspanel';
+import Statuspanel, { StatusNokkelopplysning, Statusopplysninger } from '../../components/Statuspanel';
 
 export const SendtSoknadArbeidstakerStatuspanel = ({ soknad }) => {
     return (<Statuspanel>
