+++ js/sykepengesoknad/arbeidtaker/NySoknadArbeidstaker.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import beregnSteg, { AKTIVITETER_I_SYKMELDINGSPERIODEN, FOER_DU_BEGYNNER, FRAVAER_OG_FRISKMELDING, KVITTERING, OPPSUMMERING } from '../../utils/soknad-felles/beregnSteg';
-import NyFoerDuBegynnerArbeidstakerContainer from '../../containers/sykepengesoknad-arbeidstaker-ny/NyFoerDuBegynnerArbeidstakerContainer';
-import NyFravaerOgFriskmeldingArbeidstakerContainer from '../../containers/sykepengesoknad-arbeidstaker-ny/NyFravaerOgFriskmeldingArbeidstakerContainer';
-import NyAktiviteterISykmeldingsperiodenArbeidstakerContainer from '../../containers/sykepengesoknad-arbeidstaker-ny/NyAktiviteterISykmeldingsperiodenArbeidstakerContainer';
-import NyOppsummeringArbeidstakerContainer from '../../containers/sykepengesoknad-arbeidstaker-ny/NyOppsummeringArbeidstakerContainer';
-import SoknadKvitteringSjekker from '../soknad-felles/SoknadKvitteringSjekker';
+import NyFoerDuBegynnerArbeidstakerContainer from './for-du-begynner/NyFoerDuBegynnerArbeidstakerContainer';
+import NyFravaerOgFriskmeldingArbeidstakerContainer from './fravar-og-friskmelding/NyFravaerOgFriskmeldingArbeidstakerContainer';
+import NyAktiviteterISykmeldingsperiodenArbeidstakerContainer from './aktiviteter-i-sykmeldingsperioden/NyAktiviteterISykmeldingsperiodenArbeidstakerContainer';
+import NyOppsummeringArbeidstakerContainer from './oppsummering/NyOppsummeringArbeidstakerContainer';
+import SoknadKvitteringSjekker from '../../components/soknad-felles/SoknadKvitteringSjekker';
 import { soknad as soknadPt } from '../../propTypes';
 import { KORRIGERT, NY, SENDT, UTKAST_TIL_KORRIGERING } from '../../enums/soknadstatuser';
 import NySendtSoknadArbeidstaker from './NySendtSoknadArbeidstaker';
-import Feilmelding from '../Feilmelding';
+import Feilmelding from '../../components/Feilmelding';
 
 const NySoknadArbeidstakerSkjema = (props) => {
     const { sti } = props;
