+++ js/sykepengesoknad/arbeidtaker/fravar-og-friskmelding/validerFravaerOgFriskmelding.js
@@ -1,6 +1,6 @@
-import { hentSporsmalForFravaerOgFriskmelding } from '../FravaerOgFriskmelding/FravaerOgFriskmelding';
+import { hentSporsmalForFravaerOgFriskmelding } from './FravaerOgFriskmelding';
 import validerSporsmal from '../../../validering/validerSporsmal';
-import validerFoerDuBegynner from './validerFoerDuBegynner';
+import validerFoerDuBegynner from '../for-du-begynner/validerFoerDuBegynner';
 
 export default (values, props) => {
     const sporsmal = hentSporsmalForFravaerOgFriskmelding(props.soknad);
