+++ js/sykepengesoknad/arbeidtaker/oppsummering/SoknadMottaker.js
